pyproject.toml
[tool.poetry]
name = "gallery"
version = "0.1.0"
description = ""
authors = ["DarthZiel <Studerzhan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
django-cors-headers = "^4.6.0"
pip = "^24.3.1"
django-storages = "^1.14.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
